Component Function:
TwentyThree(): This is a functional component representing the pagination example.
Inside this component, the useReducer hook is used to manage the pagination state. It initializes the state with the current page set to 1 and total items set to 0.
An array named data is generated using Array.from() to simulate a list of items. It contains 25 items labeled from 'Item 1' to 'Item 25'. This array represents the entire dataset that needs to be paginated.